Make a lot improvements to image processing Improve Measure and Arrange for Grid control Rework loading of bitmap images Make possibility for multiThreaded rendering for Game class inside UIApplication class Made some improvements to AUML parser

+++ Adamantium/Adamantium.Core/DataPointer.cs
@@ -4,20 +4,20 @@ namespace Adamantium.Core
 {
     public struct DataPointer
     {
-        public DataPointer(IntPtr pointer, int size, uint count)
+        public DataPointer(IntPtr pointer, long size, uint count)
         {
             Pointer = pointer;
             Size = size;
             Count = count;
         }
 
-        public DataPointer(IntPtr pointer, int size) : this(pointer, size, 0)
+        public DataPointer(IntPtr pointer, long size) : this(pointer, size, 0)
         {
         }
 
         public IntPtr Pointer;
 
-        public int Size;
+        public long Size;
 
         public uint Count;
