+++ Adamantium/Adamantium.Engine.Core/IIdentifiable.cs
@@ -4,6 +4,6 @@ namespace Adamantium.Engine.Core
 {
    public interface IIdentifiable
    {
-      Int64 Uid { get; }
+      UInt128 Uid { get; }
    }
 }
